fix(app): fixed style flash in tag page
