+++ src/db/server/copipes.ts
@@ -27,12 +27,39 @@ export const fetchRecentCopipes = cache(async () => {
     orderBy: { id: "desc" },
   })
   console.log("get recent copipes")
+
   const result: { id: number; title: string }[] = copipes.map((value) => {
     return {
       id: value.id,
       title: value.title!,
     }
   })
+
+  return result
+})
+
+export const fetchTagCopipes = cache(async (tagId: number, page: number) => {
+  const tagQuery = {
+    copipeToTag: {
+      some: {
+        tag_id: tagId,
+      },
+    },
+  }
+
+  const result = await prisma.$transaction([
+    prisma.copipe.findMany({
+      where: tagQuery,
+      select: copipeWithTag,
+      take: 10,
+      skip: (page - 1) * 10,
+      orderBy: { id: "desc" },
+    }),
+    prisma.copipe.count({
+      where: tagQuery,
+    }),
+  ])
+  console.log(`get copipes tagId:${tagId} page:${page}`)
 
   return result
 })
